Initial commit
Add some data structures for board structure and event types. The goal
is to have stateless events that can be recorded in a ledger. The ledger
builds up state as events are recorded, and can choose to accept or deny
any event based on a custom set of rules and the current cached state.